Move model ownership to window, add settings with plugins
This took so much effort to architect this correctly. Everything's great, but I'd like to somehow (1) get rid of the GtkOverlays, hopefully to ease the overhead, though it doesn't matter much since the widgets are reused due to the nature of ListView and (2) I hate all the weird inner types I have to make an manage, especially with the app window. Hopefully I can get rid of it without putting everything in oncecells and refcells.